style: Modify spaces, new lines and utils to restructure code
