+++ osp/models/__init__.py
@@ -1,14 +1,8 @@
 from osp.models.user_information import UserInformation
 from osp.models.form import Form
-from osp.models.utils.question import Question
-from osp.models.utils.abstract_timestamp import AbstractTimestamp
+from osp.models.question import Question
+from osp.models.abstract_timestamp import AbstractTimestamp
 from osp.models.field import (
-    Choice,
-    Dropdown,
-    Checkbox,
-    Paragraph,
-    ShortAnswer,
-    Date,
-    Time,
-    FileUpload
+    Choice, Dropdown, Checkbox, Paragraph,
+    ShortAnswer, Date, Time, FileUpload
 )
