+++ osp/models/abstract_timestamp.py
@@ -9,4 +9,5 @@ class AbstractTimestamp(models.Model):
     updated_on = models.DateTimeField(auto_now=True)
 
     class Meta:
+        
         abstract = True
