+++ osp/models/user_information.py
@@ -2,12 +2,7 @@ from django.conf import settings
 from django.contrib.auth.models import User
 from django.db import models
 
-ADM = 'admin'
-STU = 'student'
-USER = (
-    (ADM, 'Admin'),
-    (STU, 'Student'),
-)
+from osp.utils import user_types
 
 class UserInformation(models.Model):
 
@@ -18,10 +13,9 @@ class UserInformation(models.Model):
     name = models.CharField(max_length=255)
     user_type = models.CharField(
         max_length=7,
-        choices=USER,
+        choices=user_types.USERS,
         default='student'
     )
 
     def __str__(self):
-
         return f'{self.name}: {self.user_type}'
