+++ osp/utils/question_types.py
@@ -1,0 +1,18 @@
+SA = 'char'
+PA = 'text'
+CF = 'choice'
+CH = 'checkbox'
+DD = 'dropdown'
+FI = 'file'
+DA = 'date'
+TI = 'time'
+DATA_TYPES = (
+    (SA, 'Short Answer'),
+    (PA, 'Paragraph Answer'),
+    (CF, 'Choice'),
+    (CH, 'Checkbox'),
+    (DD, 'Dropdown'),
+    (FI, 'File Upload'),
+    (DA, 'Date'),
+    (TI, 'Time')
+)
