+++ osp/utils/target_user_types.py
@@ -1,0 +1,8 @@
+ALL = 'all'
+ADM = 'admin'
+STU = 'student'
+TARGET_USERS = (
+    (ALL, 'All Users'),
+    (ADM, 'Admin'),
+    (STU, 'Student'),
+)
