+++ osp/utils/user_types.py
@@ -1,0 +1,6 @@
+ADM = 'admin'
+STU = 'student'
+USERS = (
+    (ADM, 'Admin'),
+    (STU, 'Student'),
+)
